+++ HISTORY.md
@@ -1,11 +1,42 @@
 # Revision Change History
 
+## [0.7.3]
+
+Fixing a number of small bugs in preparation for upcoming releases which
+will add new features.
+
+### Additions
+
+- Added MQTT broker ID optional config input to allow the user to input the
+  MQTT broker ID the client will use (thanks @kpfleming) ([PR #220][P220])
+
+### Fixes
+
+- Increase timeout for DB Refresh and allow retry for initial request.
+  ([PR #237][P237])
+
+- Detect disconnections during poll() calls (thanks @kpfleming) ([PR 227][P227])
+
+- Modem Responder Group from Thermostat Should be 0x01 ([PR #198][P198])
+  ([Issue 154][I154])
+
+- Fixed device db find command to check the local group so multiple responsders
+  can be created. ([Issue #181][I181])
+
+- Fixed a bug in the modem database class when removing an entry (thanks
+  @krkeegan) ([PR#196][P196])
+
+- Changed the MQTT Remote to never mark messages for retain so the broker
+  doesn't get out of sync with the device. ([Issue #I210][I210])
+
+
 ## [0.7.2]
 
 ### Fixes
 
 - Fixed an issue causing 100% cpu usage introduced in the scene sync code.
   ([Issue #195)[I195])
+
 
 ## [0.7.1]
 
@@ -363,3 +394,10 @@
 [I192]: https://github.com/TD22057/insteon-mqtt/issues/192
 [I193]: https://github.com/TD22057/insteon-mqtt/issues/193
 [I195]: https://github.com/TD22057/insteon-mqtt/issues/195
+[P196]: https://github.com/TD22057/insteon-mqtt/pull/196
+[I210]: https://github.com/TD22057/insteon-mqtt/issues/210
+[P220]: https://github.com/TD22057/insteon-mqtt/pull/220
+[I181]: https://github.com/TD22057/insteon-mqtt/issues/181
+[I154]: https://github.com/TD22057/insteon-mqtt/issues/154
+[P227]: https://github.com/TD22057/insteon-mqtt/pull/227
+[P237]: https://github.com/TD22057/insteon-mqtt/pull/227